Rename pages to screens folder
