+++ LabProject/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Home from './src/pages/Home';
+import Home from './src/screens/Home';
 import { ThemeProvider } from 'react-native-elements';
 import theme from './src/theme/theme';
