+++ LabProject/package.json
@@ -11,6 +11,7 @@
         "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
     },
     "dependencies": {
+        "@react-navigation/bottom-tabs": "^6.0.7",
         "@react-navigation/native": "^6.0.4",
         "@react-navigation/native-stack": "^6.2.2",
         "react": "17.0.2",
